Had to fix family combobox again. It needed its own collection
In the middle of implementing some "Collection Changed" changes I
noticed a bug in "cbFamilyGrid" where, if you add an item to the bottom
of the FAMILY grid list, It adds as ID 0,  (puts itself in the front of
the combobox) then the DB updates the ID to its permanent ID (so it
reorders itself to the end of the combobox) and this process shifted
every FAMILYID on the children tab up by  +1.  So I fixed that by
creating a new collection, which is only updated AFTER the item has been
Saved to the DB (receives a permanent ID at the end of the list). Then I
noticed another problem if you delete a record out of the middle, it
would cause 2 bugs. The first was, it would set the item you deleted to
a -1 CHILDRENS.FAMILYID which will crash if you try to save, because
thats not a valid foreignkey (mapped as FAMILIES.ID). The second bug was
it would change all the NAMES in the the cbFamilygrid, subsequent to the
index you deleted. (all the IDs will remain valid, except the one you
deleted = the one you deleted will show as valid until you click it,
upon which the ID and the name will go blank)... All of these were
solved by changing "SelectedIndex" to "SelectedValue" in cbFamilyGrid.
I will clean up the files after I'm sure.