fix updates (INotify), bindings.
bind to observablecollections mostly get rid of collectionviewsource.
Can set the observablecollection to be sorted on view by using an
internal property (no need for a seperate variable/identifier name).
Comment out the old viewsources.  Add 1 button for each tab to select
the last row (Currently in Codebehind). Set the
"IsSynchronizedWithCurrentItem=True" on the datagrids with the
rowdetails so you can see them in the designer while editing to make it
easier.  Change Family combobox to show firstname when picking the right
one to make it easier.  City ComboBox is now binded to
"ElementName=customChromeWindow,Path=DataContext.CitiesCollection" (this
is how to reference the main scope when inside a datagrid (rowdetails
count as inside a datagrid). The composite collections
<CollectionContainer  need to use "Source={x:Reference " instead of
ElementName. Binded the Civility textbox inside the Conjoint & Payeur
rowdetail tabs.